Collapse Dataset inheritance.
